requirements.txt
absl-py==2.1.0
altair==5.5.0
anyio==4.4.0
apyori==1.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.11.1
bleach==6.1.0
blinker==1.8.2
boto3==1.35.49
botocore==1.35.49
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
dynaconf==3.2.6
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.16.1
Flask==3.0.3
Flask-MySQLdb==2.0.0
flatbuffers==24.3.25
fonttools==4.52.4
fpdf==1.7.2
fqdn==1.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
hvac==2.3.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keras==3.4.1
kiwisolver==1.4.5
knox==0.1.14
libclang==18.1.1
loguru==0.7.2
lxml==4.9.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.22.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.0
plumbum==1.9.0
ply==3.11
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.1.0
shortuuid==1.0.13
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.41.1
sympy==1.12.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
validators==0.34.0
virtualenv==20.27.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
win32-setctime==1.1.0
wrapt==1.16.0
